Moves Config directory finding to FEXCore
This will be necessary in the next commit
